backend/service/requirements.txt
flask
pymongo
requests
requests-html
python-dotenv
flask_cors